# openapi/api-v1.yaml
openapi: 3.0.3
info:
  title: EsquiloSpeak API
  description: API specifications for the EsquiloSpeak MVP language learning app.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Local API Gateway
security:
  - BearerAuth: []
paths:
...
  /courses/{courseId}/attempts:
    post:
      summary: Submit question attempt
      description: Records user answer to a question in a lesson for progress and review scheduler.
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: string
          example: "en_for_vi"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AttemptSubmitRequest'
      responses:
        '200':
          description: Attempt processed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  isCorrect:
                    type: boolean
                    example: true
                  correctAnswer:
                    type: string
                    example: "Xin chào"
                  explanation:
                    type: string
                    example: "'Hello' dịch sang tiếng Việt nghĩa là 'Xin chào'."
                  reviewCreated:
                    type: boolean
                    example: false
        '400':
          $ref: '#/components/responses/ErrorResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /sync/attempts:
    post:
      summary: Sync offline attempts
      description: Uploads a list of cached offline question attempts to reconcile progress.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - attempts
              properties:
                attempts:
                  type: array
                  items:
                    $ref: '#/components/schemas/AttemptSubmitRequest'
      responses:
        '200':
          description: Offline progress synchronized successfully
          content:
            application/json:
              schema:
                type: object
                required:
                  - success
                  - syncedCount
                  - results
                properties:
                  success:
                    type: boolean
                    example: true
                  syncedCount:
                    type: integer
                    format: int32
                    example: 3
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SyncResult'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
components:
  schemas:
    AttemptSubmitRequest:
      type: object
      required:
        - clientRequestId
        - deviceId
        - courseId
        - sourceLanguage
        - targetLanguage
        - lessonId
        - lessonVersionId
        - questionId
        - questionVersionId
        - selectedAnswer
        - responseTimeMs
        - answeredAt
      properties:
        clientRequestId:
          type: string
          example: "crq_001"
        deviceId:
          type: string
          example: "device_abc"
        courseId:
          type: string
          example: "en_for_vi"
        sourceLanguage:
          type: string
          example: "vi"
        targetLanguage:
          type: string
          example: "en"
        lessonId:
          type: string
          example: "lesson_101"
        lessonVersionId:
          type: string
          example: "lesson_101_v1"
        questionId:
          type: string
          example: "q_201"
        questionVersionId:
          type: string
          example: "q_201_v1"
        selectedAnswer:
          type: string
          example: "Xin chào"
        responseTimeMs:
          type: integer
          example: 4200
        usedHint:
          type: boolean
          example: false
        attemptSource:
          type: string
          enum: [lesson, review]
          example: "lesson"
        answeredAt:
          type: string
          example: "2026-06-23T10:00:00Z"
    SyncResult:
      type: object
      required:
        - clientRequestId
        - status
      properties:
        clientRequestId:
          type: string
          example: "crq_001"
        status:
          type: string
          enum: [SYNCED, DUPLICATE, FAILED]
          example: "SYNCED"
        errorCode:
          type: string
          nullable: true
          example: "QUESTION_NOT_FOUND"
        message:
          type: string
          nullable: true
          example: "Question not found in content-service"
  responses:
    ErrorResponse:
      description: Standard Error Response
      content:
        application/json:
          schema:
            type: object
            required:
              - error
              - meta
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: "BAD_REQUEST"
                  message:
                    type: string
                    example: "Invalid input values"
                  details:
                    type: object
                    example: {}
              meta:
                type: object
                required:
                  - requestId
                  - apiVersion
                properties:
                  requestId:
                    type: string
                    example: "req_xyz987"
                  apiVersion:
                    type: string
                    example: "v1"
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            required:
              - error
              - meta
            properties:
              error:
                type: object
                required:
                  - code
                  - message
                properties:
                  code:
                    type: string
                    example: "UNAUTHORIZED"
                  message:
                    type: string
                    example: "Token is expired or invalid"
              meta:
                type: object
                required:
                  - requestId
                  - apiVersion
                properties:
                  requestId:
                    type: string
                    example: "req_xyz987"
                  apiVersion:
                    type: string
                    example: "v1"
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Nhập JWT access token để gọi các API được bảo vệ.